adding an access check method: checkPermission, to the AdServiceImpl …

+++ src/main/java/com/skypro/adsonline/service/impl/AdServiceImpl.java
@@ -1,20 +1,21 @@
 package com.skypro.adsonline.service.impl;
 
-import com.skypro.adsonline.dto.Ads;
-import com.skypro.adsonline.dto.CreateAds;
-import com.skypro.adsonline.dto.FullAds;
-import com.skypro.adsonline.dto.ResponseWrapperAds;
+import com.skypro.adsonline.dto.*;
 import com.skypro.adsonline.exception.AdNotFoundException;
 import com.skypro.adsonline.model.AdEntity;
 import com.skypro.adsonline.repository.AdRepository;
 import com.skypro.adsonline.security.SecurityUser;
 import com.skypro.adsonline.service.AdService;
 import com.skypro.adsonline.utils.AdMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
+import static com.skypro.adsonline.constant.ErrorMessage.ACCESS_DENIED_MSG;
 import static com.skypro.adsonline.constant.ErrorMessage.AD_NOT_FOUND_MSG;
 
 @Service
@@ -42,7 +43,8 @@ public class AdServiceImpl implements AdService {
     }
 
     @Override
-    public boolean removeAd(Integer id) {
+    public boolean removeAd(Integer id, SecurityUser currentUser) {
+        checkPermission(id, currentUser);
         AdEntity ad = adRepository.findById(id).orElseThrow(() -> new AdNotFoundException(AD_NOT_FOUND_MSG.formatted(id)));
         adRepository.delete(ad);
         return true;
@@ -78,8 +80,10 @@ public class AdServiceImpl implements AdService {
     }
 
     @Override
-    public Ads updateAds(Integer id, CreateAds ads) {
-        AdEntity ad = adRepository.findById(id).orElseThrow(() -> new AdNotFoundException(AD_NOT_FOUND_MSG.formatted(id)));
+    public Ads updateAds(Integer id, CreateAds ads, SecurityUser currentUser) {
+        checkPermission(id, currentUser);
+        AdEntity ad = adRepository.findById(id).orElseThrow(() ->
+                new AdNotFoundException(AD_NOT_FOUND_MSG.formatted(id)));
 
         ad.setTitle(ads.getTitle());
         ad.setDescription(ads.getDescription());
@@ -87,5 +91,19 @@ public class AdServiceImpl implements AdService {
 
         adRepository.save(ad);
         return adMapper.mapToAdDto(ad);
+    }
+
+    private void checkPermission(Integer adId, SecurityUser currentUser) {
+        Integer authorId = adRepository
+                .findById(adId)
+                .orElseThrow(() ->
+                        new AdNotFoundException(AD_NOT_FOUND_MSG.formatted(adId)))
+                .getAuthor()
+                .getId();
+        if (!authorId.equals(currentUser.getUser().getId())
+                && !currentUser.getAuthorities().
+                contains(new SimpleGrantedAuthority(Role.ADMIN.name()))) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, ACCESS_DENIED_MSG);
+        }
     }
 }
